+++ providers/microsoft/azure/docs/changelog.rst
@@ -49,6 +49,7 @@ Bug Fixes
 Features
 ~~~~~~~~
 
+* ``Add Azure AI Foundry Hosted agents operators``
 * ``Add GCSToAzureBlobStorageOperator for GCS to Azure Blob transfer (#64966)``
 
 Bug Fixes